+++ CHANGELOG.md
@@ -16,3 +16,6 @@
 ## 2026-02-10T03:22:27.183977+00:00
 - Workflows API endpoints: workflow-marketplace_patch_2026-02-10T03-22-18.260501+00-00.diff
 
+## 2026-02-10T03:26:57.593466+00:00
+- Workflows API endpoints: workflow-marketplace_patch_2026-02-10T03-26-48.717012+00-00.diff
+